Fix scroll rate being broken at high FPS

+++ src/TSMapEditor/UI/MapUI.cs
@@ -603,7 +603,7 @@ namespace TSMapEditor.UI
             // Make scroll rate independent of FPS
             // Scroll rate is designed for 60 FPS
             // 1000 ms (1 second) divided by 60 frames =~ 16.667 ms / frame
-            int scrollRate = (int)(this.scrollRate * (gameTime.ElapsedGameTime.TotalMilliseconds / 16.667));
+            float scrollRate = (float)(this.scrollRate * (gameTime.ElapsedGameTime.TotalMilliseconds / 16.667));
 
             if (IsActive)
             {
